+++ src/OpenTelemetry.Instrumentation.SqlClient/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+* Renamed from `AddSqlClientDependencyInstrumentation` to `AddSqlClientInstrumentation`
+
 ## 0.3.0-beta
 
 Released 2020-07-23